done, next deploy

+++ client/src/components/main-section/Task.vue
@@ -1,54 +1,59 @@
 <template>
   <div>
-    <div >
+    <div>
       <div class="task-section">
         <h3 style="text-transform: capitalize;font-size:1.2em;">{{task.title}}</h3>
         <p>{{task.description}}</p>
         <span>
-          <i class="fas fa-edit edit-button" v-on:click="$emit('emitEdit')" ></i>
+          <i class="fas fa-edit edit-button" v-on:click="$emit('emitEdit',task)"></i>
         </span>
         <span>
           <i class="fas fa-trash-alt delete-button" v-on:click="deleteTask"></i>
         </span>
-      </div>
-      <div v-if="editStatus">
-        <EditForm></EditForm>
       </div>
     </div>
   </div>
 </template>
 
 <script>
-import EditForm from './EditForm'
-
 export default {
   name: "Task",
-  components:{
-    EditForm
-  },
-  data(){
-    return {
-      editStatus : false
-    }
-  },
-  props:["task"],
-  methods:{
-    deleteTask: function(){
-      let self = this 
-      axios({
-        method:'delete',
-        url:`http://localhost:3000/tasks/${this.task.id}`,
-        headers:{
-          access_token: localStorage.getItem('access_token')
+  props: ["task"],
+  methods: {
+    deleteTask: function() {
+      let self = this;
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!"
+      }).then(result => {
+        if (result.value) {
+          Swal.fire("Deleted!", "Your file has been deleted.", "success");
+          axios({
+            method: "delete",
+            url: `http://localhost:3000/tasks/${this.task.id}`,
+            headers: {
+              access_token: localStorage.getItem("access_token")
+            }
+          })
+            .then(function(response) {
+              self.$emit("emitDelete");
+            })
+            .catch(function(error) {
+              Swal.fire({
+                icon: "info",
+                title: "Unauthorized",
+                text: "You can not delete this data!",
+                footer: "<a href>Why do I have this issue?</a>"
+              });
+              console.log(error.response);
+            });
         }
-      })
-      .then(function(response){
-        console.log(response)
-        self.$emit('emitDelete')
-      })
-      .catch(function(error){
-        console.log(error.response)
-      })
+      });
     }
   }
 };
